added route & service methods to process buckydrop orders severally
