+++ hamza-server/src/api/custom/config/route.ts
@@ -3,7 +3,7 @@ import { RouteHandler } from '../../route-handler';
 import { Config } from '../../../config';
 
 /*This route does one thing: it updates the checkoutMode property when CartCompletionStrategy runs, based off the
- * environment variable CHECKOUT_MODE. This is a simple GET request that returns the payment mode. */
+ * environment variable     . This is a simple GET request that returns the payment mode. */
 
 export const GET = async (req: MedusaRequest, res: MedusaResponse) => {
     const handler: RouteHandler = new RouteHandler(
